// taiff_web/src/app/interfaces/folhaDeRostoIterface.ts
export type formFolhaDeRosto = {
    nomeDoTeste?: string;
    executor?: string;
    tipo?: string;
    fabricante?: string;
    modeloEmPlaca?: string;
    paisFabricado?: string;
    modeloComercial?: string;
    frequencia?: number;
    tensao?: number;
    potencia?: number;
    local?: string;
}